+++ ChangeLog.md
@@ -13,7 +13,7 @@ and this project adheres to
 
 ### Added
 
-- ...
+- function `str/levenshtein`
 
 
 ## [1.10.0] - 2021-10-01